requirements.txt
aiohttp==1.0.5
async-timeout==1.2.0
cffi==1.9.1
chardet==2.3.0
discord.py==0.16.7
multidict==2.1.4
names==0.3.0
pendulum==1.2.0
pony==0.7.1
pycparser==2.17
PyNaCl==1.0.1
python-dateutil==2.6.0
pytz==2016.10
pytzdata==2017.2
six==1.10.0
tzlocal==1.3
websockets==3.2
